requirements.txt
watchdog
requests
python-dotenv
re
time
collections
os
csv
json
datetime
